refactor(KakaoMap): 코드 정리
